+++ nautilus_core/core/src/python/uuid.rs
@@ -1,0 +1,103 @@
+// -------------------------------------------------------------------------------------------------
+//  Copyright (C) 2015-2023 Nautech Systems Pty Ltd. All rights reserved.
+//  https://nautechsystems.io
+//
+//  Licensed under the GNU Lesser General Public License Version 3.0 (the "License");
+//  You may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at https://www.gnu.org/licenses/lgpl-3.0.en.html
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+// -------------------------------------------------------------------------------------------------
+
+use std::{
+    collections::hash_map::DefaultHasher,
+    hash::{Hash, Hasher},
+    str::FromStr,
+};
+
+use pyo3::{
+    prelude::*,
+    pyclass::CompareOp,
+    types::{PyBytes, PyTuple},
+};
+
+use super::to_pyvalue_err;
+use crate::uuid::UUID4;
+
+#[pymethods]
+impl UUID4 {
+    #[new]
+    fn py_new(value: Option<&str>) -> PyResult<Self> {
+        match value {
+            Some(val) => Self::from_str(val).map_err(to_pyvalue_err),
+            None => Ok(Self::new()),
+        }
+    }
+
+    fn __setstate__(&mut self, py: Python, state: PyObject) -> PyResult<()> {
+        let bytes: &PyBytes = state.extract(py)?;
+        let slice = bytes.as_bytes();
+
+        if slice.len() != 37 {
+            return Err(to_pyvalue_err(
+                "Invalid state for deserialzing, incorrect bytes length",
+            ));
+        }
+
+        self.value.copy_from_slice(slice);
+        Ok(())
+    }
+
+    fn __getstate__(&self, _py: Python) -> PyResult<PyObject> {
+        Ok(PyBytes::new(_py, &self.value).to_object(_py))
+    }
+
+    fn __reduce__(&self, py: Python) -> PyResult<PyObject> {
+        let safe_constructor = py.get_type::<Self>().getattr("_safe_constructor")?;
+        let state = self.__getstate__(py)?;
+        Ok((safe_constructor, PyTuple::empty(py), state).to_object(py))
+    }
+
+    #[staticmethod]
+    fn _safe_constructor() -> PyResult<Self> {
+        Ok(Self::new()) // Safe default
+    }
+
+    fn __richcmp__(&self, other: &Self, op: CompareOp, py: Python<'_>) -> Py<PyAny> {
+        match op {
+            CompareOp::Eq => self.eq(other).into_py(py),
+            CompareOp::Ne => self.ne(other).into_py(py),
+            _ => py.NotImplemented(),
+        }
+    }
+
+    fn __hash__(&self) -> isize {
+        let mut h = DefaultHasher::new();
+        self.hash(&mut h);
+        h.finish() as isize
+    }
+
+    fn __str__(&self) -> String {
+        self.to_string()
+    }
+
+    fn __repr__(&self) -> String {
+        format!("{}('{}')", stringify!(UUID4), self)
+    }
+
+    #[getter]
+    #[pyo3(name = "value")]
+    fn py_value(&self) -> String {
+        self.to_string()
+    }
+
+    #[staticmethod]
+    #[pyo3(name = "from_str")]
+    fn py_from_str(value: &str) -> PyResult<Self> {
+        Self::from_str(value).map_err(to_pyvalue_err)
+    }
+}
