+++ nautilus_core/model/src/data/trade_py.rs
@@ -20,8 +20,8 @@ use std::{
 };
 
 use nautilus_core::{
-    python::to_pyvalue_err,
-    serialization::{from_dict_pyo3, Serializable},
+    python::{serialization::from_dict_pyo3, to_pyvalue_err},
+    serialization::Serializable,
     time::UnixNanos,
 };
 use pyo3::{
