+++ nautilus_core/model/src/instruments/synthetic_api.rs
@@ -19,7 +19,7 @@ use std::{
 };
 
 use nautilus_core::{
-    cvec::CVec,
+    ffi::cvec::CVec,
     parsing::{bytes_to_string_vec, string_vec_to_bytes},
     string::{cstr_to_string, str_to_cstr},
     time::UnixNanos,
