+++ nautilus_trader/core/rust/core.pxd
@@ -24,10 +24,6 @@ cdef extern from "../includes/core.h":
     cdef struct UUID4_t:
         uint8_t value[37];
 
-    void cvec_drop(CVec cvec);
-
-    CVec cvec_new();
-
     # Converts seconds to nanoseconds (ns).
     uint64_t secs_to_nanos(double secs);
 
@@ -48,6 +44,9 @@ cdef extern from "../includes/core.h":
 
     # Converts nanoseconds (ns) to microseconds (μs).
     uint64_t nanos_to_micros(uint64_t nanos);
+
+    # Converts a UNIX nanoseconds timestamp to an ISO 8601 formatted C string pointer.
+    const char *unix_nanos_to_iso8601_cstr(uint64_t timestamp_ns);
 
     # Return the decimal precision inferred from the given C string.
     #
@@ -82,6 +81,10 @@ cdef extern from "../includes/core.h":
 
     # Returns the current nanoseconds since the UNIX epoch.
     uint64_t unix_timestamp_ns();
+
+    void cvec_drop(CVec cvec);
+
+    CVec cvec_new();
 
     UUID4_t uuid4_new();
